move behaviour to p2p mod and split into sub-mods
