+++ src/lib.rs
@@ -13,7 +13,6 @@ use rocket::{fairing::AdHoc, figment::Figment, http::Header, Build, Rocket};
 pub mod allow_list;
 pub mod auth_guards;
 pub mod authorization;
-pub mod behaviour;
 pub mod capabilities;
 pub mod cas;
 pub mod config;
@@ -21,6 +20,7 @@ pub mod indexes;
 pub mod kv;
 pub mod manifest;
 pub mod orbit;
+pub mod p2p;
 pub mod prometheus;
 pub mod relay;
 pub mod routes;
